fix: room api not using next app router
